auth module finish

+++ src/modules/books/dto/register-book.input.ts
@@ -1,4 +1,4 @@
-import { Field, InputType, Int } from '@nestjs/graphql';
+import { Field, Float, InputType, Int } from '@nestjs/graphql';
 import { IsEnum } from 'class-validator';
 import { BookCoverTypes } from 'src/common/enums/book-cover-types';
 import { BookTitleTypes } from 'src/common/enums/book-title-types';
@@ -13,7 +13,7 @@ export class RegisterBookInput {
     nullable: false,
   })
   bookTitle: string;
-  @Field(() => Int, {
+  @Field(() => Float, {
     description: 'Publish date in epoch milliseconds',
     nullable: false,
   })
@@ -22,7 +22,7 @@ export class RegisterBookInput {
     nullable: false,
   })
   editionNumber: number;
-  @Field(() => Int, {
+  @Field(() => Float, {
     description: 'Edition date in epoch milliseconds',
     nullable: false,
   })
